Fix compile errors post-PR20
PR-20 introduced build errors when not compiling using the IBM BB API.
This patch fixes the issue.
